refactor(encoders): consolidate into xr_toolz.transforms.encoders (Epic F5)
Closes #76, #82, #83, #84.

Move the 12 encoder Layer-0 functions from xr_toolz.geo._src.encoders to
xr_toolz.transforms._src.encoders/, organized by what they encode per D8:

  coord_space.py — lon_360_to_180, lon_180_to_360, lat_180_to_90, lat_90_to_180
  coord_time.py  — time_rescale, time_unrescale, encode_time_cyclical, encode_time_ordinal
  basis.py       — cyclical_encode, fourier_features, random_fourier_features, positional_encoding

Public surface lives at xr_toolz.transforms.encoders (subpackage), with
both flat and per-submodule import paths working:

  from xr_toolz.transforms.encoders import cyclical_encode
  from xr_toolz.transforms.encoders.coord_space import lon_360_to_180

F5.2 — audited transforms/operators.py: no encoder Tier C operators
exist today, so nothing to retarget. Filed follow-up #95 to add
CyclicalEncode / FourierFeatures / PositionalEncoding wrappers.

F5.3 — geo/__init__.py drops direct encoder imports and routes the 12
names through PEP-562 __getattr__ with a DeprecationWarning pointing at
the new home (mirrors the F1 metrics pattern). Internal callsite in
geo/_src/validation.py imports from the new home directly.
geo/_src/encoders.py is kept as a deep-import shim so existing
``from xr_toolz.geo._src.encoders import …`` callers continue to
resolve for one release.

CHANGELOG: Unreleased / Deprecated entry covering all 12 names.

Tests: +18 in tests/test_encoder_imports.py covering the canonical
surface, submodule paths, identity equivalence, the parametrized
deprecation surface (12 names), and silent plain-import.

Suite: 545 → 563 (+18). Lint, format, ty, mkdocs strict net-zero new
warnings.
