+++ src/xr_toolz/transforms/encoders/__init__.py
@@ -1,0 +1,46 @@
+"""Public encoder surface — coord-space, coord-time, and basis encodings.
+
+Re-exports the pure functions from :mod:`xr_toolz.transforms._src.encoders`.
+Submodules :mod:`coord_space`, :mod:`coord_time`, and :mod:`basis` are
+also importable as ``xr_toolz.transforms.encoders.<sub>`` for callers
+who want a narrower import.
+"""
+
+from xr_toolz.transforms.encoders import basis, coord_space, coord_time
+from xr_toolz.transforms.encoders.basis import (
+    cyclical_encode,
+    fourier_features,
+    positional_encoding,
+    random_fourier_features,
+)
+from xr_toolz.transforms.encoders.coord_space import (
+    lat_90_to_180,
+    lat_180_to_90,
+    lon_180_to_360,
+    lon_360_to_180,
+)
+from xr_toolz.transforms.encoders.coord_time import (
+    encode_time_cyclical,
+    encode_time_ordinal,
+    time_rescale,
+    time_unrescale,
+)
+
+
+__all__ = [
+    "basis",
+    "coord_space",
+    "coord_time",
+    "cyclical_encode",
+    "encode_time_cyclical",
+    "encode_time_ordinal",
+    "fourier_features",
+    "lat_90_to_180",
+    "lat_180_to_90",
+    "lon_180_to_360",
+    "lon_360_to_180",
+    "positional_encoding",
+    "random_fourier_features",
+    "time_rescale",
+    "time_unrescale",
+]
