+++ src/xr_toolz/transforms/encoders/basis.py
@@ -1,0 +1,16 @@
+"""Public re-export of basis encoders."""
+
+from xr_toolz.transforms._src.encoders.basis import (
+    cyclical_encode,
+    fourier_features,
+    positional_encoding,
+    random_fourier_features,
+)
+
+
+__all__ = [
+    "cyclical_encode",
+    "fourier_features",
+    "positional_encoding",
+    "random_fourier_features",
+]
